# docs/swagger.yaml
definitions:
...
  domain.Section:
    properties:
      current_capacity:
        type: integer
      current_temperature:
        type: integer
      maximum_capacity:
        type: integer
      minimum_capacity:
        type: integer
      minimum_temperature:
        type: integer
      product_type_id:
        type: integer
      section_id:
        type: integer
      section_number:
        type: integer
      warehouse_id:
        type: integer
    required:
    - current_capacity
    - current_temperature
    - maximum_capacity
    - minimum_capacity
    - minimum_temperature
    - product_type_id
    - section_number
    - warehouse_id
    type: object
  domain.SectionReportProducts:
    properties:
      product_count:
        type: integer
      section_id:
        description: Section
        type: integer
      section_number:
        type: integer
    type: object
...
  web.errorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  web.response:
    properties:
      data: {}
    type: object
info:
  contact: {}
  description: This API was created as part of Mercado Libre's Go backend bootcamp
    for training purposes.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es-ar/terminos-y-condiciones
  title: MeLi Bootcamp API
  version: "1.0"
paths:
...
  /api/v1/sections:
    get:
      description: Get All Sections
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: List sections
      tags:
      - Sections
    post:
      consumes:
      - application/json
      description: Create section
      parameters:
      - description: Section to create
        in: body
        name: section
        required: true
        schema:
          $ref: '#/definitions/domain.Section'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Create section
      tags:
      - Sections
  /api/v1/sections/{id}:
    delete:
      description: Delete section
      parameters:
      - description: section id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Delete section
      tags:
      - Sections
    get:
      description: Get Section by Id
      parameters:
      - description: section id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Section by id
      tags:
      - Sections
    patch:
      consumes:
      - application/json
      description: Update section
      parameters:
      - description: section id
        in: path
        name: id
        required: true
        type: integer
      - description: Section to update
        in: body
        name: section
        required: true
        schema:
          $ref: '#/definitions/domain.Section'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Update section
      tags:
      - Sections
  /api/v1/sections/reportProducts:
    get:
      description: Get the quantity of products of each section or the quantity of
        products for a determined section
      parameters:
      - description: Section id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.SectionReportProducts'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Report Products
      tags:
      - Sections
...
securityDefinitions:
  ApiKeyAuth:
    description: start with Bearer
    in: header
    name: x-tiger-token
    type: apiKey
swagger: "2.0"